refactor(pii): split lender-vocabulary resolution out of pii_redaction
Adding the LTV-basis guard pushed pii_redaction.py to 915 lines, past the
900-line monolith gate. Splitting is the repo's established answer there,
not an allowlist entry.

"Which public alias does this raw servicer string map to" is a separate
concern from "which fields may leave a repository", so the lender
dictionary, LenderRefResolver, and its process-wide accessors move to
backend/services/lender_ref.py. pii_redaction re-exports every name, so no
caller or test import path changes -- the split is a file boundary, not an
API change. 915 -> 732 lines.
